Split the abilities into event and RSVP domain classes
Replaces the single Core\Abilities class with Event\Abilities and
Rsvp\Abilities, each loaded by its own subsystem Setup, so the two
supports stay independent:

- gatherpress/get-event-rsvp-counts becomes gatherpress/get-rsvp-counts
  and takes a post_id, since RSVPs are their own post type support and
  need not belong to an event.
- The RSVP permission callback gates on gatherpress-rsvp support rather
  than gatherpress-event-date, by way of Event::can_read_rsvps().
- The upcoming events listing drops its read gate. WP_Query already
  limits it to what the viewer could read on the event archive, so an
  anonymous visitor can list what the archive shows them.
- The shared ability category is registered by whichever subsystem gets
  there first, so either class can load on its own.
